requirements.txt
biopython
gtfparse
matplotlib
seaborn
pandas
numpy
scipy
matplotlib_venn
upsetplot
openpyxl